+++ R/zzz.R
@@ -60,6 +60,9 @@
   }
 
   registerClusterTypes()
+
+  ## R CMD build
+  register_vignette_engine_during_build_only(pkgname)
 }